+++ crates/conversions/src/byte_array.rs
@@ -1,6 +1,5 @@
 use cairo_felt::Felt252;
 use cairo_lang_utils::byte_array::{BYTES_IN_WORD, BYTE_ARRAY_MAGIC};
-use itertools::chain;
 use num_traits::Num;
 
 pub struct ByteArray {
@@ -9,8 +8,8 @@ pub struct ByteArray {
     pending_word: Felt252,
 }
 
-impl From<String> for ByteArray {
-    fn from(value: String) -> Self {
+impl From<&str> for ByteArray {
+    fn from(value: &str) -> Self {
         let chunks = value.as_bytes().chunks_exact(BYTES_IN_WORD);
         let remainder = chunks.remainder();
         let pending_word_len = remainder.len();
@@ -29,14 +28,14 @@ impl From<String> for ByteArray {
 impl ByteArray {
     #[must_use]
     pub fn serialize(self) -> Vec<Felt252> {
-        chain!(
-            [
-                Felt252::from_str_radix(BYTE_ARRAY_MAGIC, 16).unwrap(),
-                self.words.len().into()
-            ],
-            self.words.into_iter(),
-            [self.pending_word, self.pending_word_len.into()]
-        )
-        .collect()
+        let mut result = Vec::with_capacity(self.words.len() + 4);
+
+        result.push(Felt252::from_str_radix(BYTE_ARRAY_MAGIC, 16).unwrap());
+        result.push(self.words.len().into());
+        result.extend(self.words);
+        result.push(self.pending_word);
+        result.push(self.pending_word_len.into());
+
+        result
     }
 }