+++ crates/forge/tests/integration/deploy.rs
@@ -8,14 +8,14 @@ use test_utils::{assert_passed, test_case};
 fn error_handling() {
     let test = test_case!(
         indoc!(
-            r"
+            r#"
         use result::ResultTrait;
         use snforge_std::{ declare, ContractClass, ContractClassTrait };
         use array::ArrayTrait;
 
         #[test]
         fn error_handling() {
-            let contract = declare('PanickingConstructor');
+            let contract = declare("PanickingConstructor");
 
             match contract.deploy(@ArrayTrait::new()) {
                 Result::Ok(_) => panic_with_felt252('Should have panicked'),
@@ -25,7 +25,7 @@ fn error_handling() {
                 }
             }
         }
-    "
+    "#
         ),
         Contract::from_code_path(
             "PanickingConstructor".to_string(),
@@ -43,7 +43,7 @@ fn error_handling() {
 fn deploy_syscall_check() {
     let test = test_case!(
         indoc!(
-            r"
+            r#"
         use snforge_std::{declare, test_address};
         use starknet::{SyscallResult, deploy_syscall};
         
@@ -55,7 +55,7 @@ fn deploy_syscall_check() {
 
         #[test]
         fn deploy_syscall_check() {
-            let contract = declare('DeployChecker');
+            let contract = declare("DeployChecker");
             let salt = 1;
             let calldata = array![10];
         
@@ -69,7 +69,7 @@ fn deploy_syscall_check() {
             let (contract_address_from_zero, _) = deploy_syscall(contract.class_hash, salt, calldata.span(), true).unwrap();
             assert(contract_address != contract_address_from_zero, 'deploy from zero no effect');
         }
-    "
+    "#
         ),
         Contract::from_code_path(
             "DeployChecker".to_string(),