+++ crates/forge/tests/integration/roll.rs
@@ -8,7 +8,7 @@ use test_utils::{assert_passed, test_case};
 fn roll_basic() {
     let test = test_case!(
         indoc!(
-            r"
+            r#"
             use result::ResultTrait;
             use array::ArrayTrait;
             use option::OptionTrait;
@@ -24,7 +24,7 @@ fn roll_basic() {
 
             #[test]
             fn test_stop_roll() {
-                let contract = declare('RollChecker');
+                let contract = declare("RollChecker");
                 let contract_address = contract.deploy(@ArrayTrait::new()).unwrap();
                 let dispatcher = IRollCheckerDispatcher { contract_address };
 
@@ -43,7 +43,7 @@ fn roll_basic() {
 
             #[test]
             fn test_roll_multiple() {
-                let contract = declare('RollChecker');
+                let contract = declare("RollChecker");
 
                 let roll_checker1 = IRollCheckerDispatcher { contract_address: contract.deploy(@ArrayTrait::new()).unwrap() };
                 let roll_checker2 = IRollCheckerDispatcher { contract_address: contract.deploy(@ArrayTrait::new()).unwrap() };
@@ -70,7 +70,7 @@ fn roll_basic() {
 
             #[test]
             fn test_roll_all() {
-                let contract = declare('RollChecker');
+                let contract = declare("RollChecker");
 
                 let roll_checker1 = IRollCheckerDispatcher { contract_address: contract.deploy(@ArrayTrait::new()).unwrap() };
                 let roll_checker2 = IRollCheckerDispatcher { contract_address: contract.deploy(@ArrayTrait::new()).unwrap() };
@@ -97,7 +97,7 @@ fn roll_basic() {
 
             #[test]
             fn test_roll_all_stop_one() {
-                let contract = declare('RollChecker');
+                let contract = declare("RollChecker");
                 let contract_address = contract.deploy(@ArrayTrait::new()).unwrap();
                 let dispatcher = IRollCheckerDispatcher { contract_address };
 
@@ -113,7 +113,7 @@ fn roll_basic() {
                 let new_block_number = dispatcher.get_block_number();
                 assert(new_block_number == old_block_number, 'Block num did not change back');
             }
-        "
+        "#
         ),
         Contract::from_code_path(
             "RollChecker".to_string(),
@@ -131,7 +131,7 @@ fn roll_basic() {
 fn roll_complex() {
     let test = test_case!(
         indoc!(
-            r"
+            r#"
             use result::ResultTrait;
             use array::ArrayTrait;
             use option::OptionTrait;
@@ -146,14 +146,14 @@ fn roll_complex() {
             }
 
             fn deploy_roll_checker()  -> IRollCheckerDispatcher {
-                let contract = declare('RollChecker');
+                let contract = declare("RollChecker");
                 let contract_address = contract.deploy(@ArrayTrait::new()).unwrap();
                 IRollCheckerDispatcher { contract_address }
             }
 
             #[test]
             fn roll_complex() {
-                let contract = declare('RollChecker');
+                let contract = declare("RollChecker");
 
                 let roll_checker1 = IRollCheckerDispatcher { contract_address: contract.deploy(@ArrayTrait::new()).unwrap() };
                 let roll_checker2 = IRollCheckerDispatcher { contract_address: contract.deploy(@ArrayTrait::new()).unwrap() };
@@ -192,7 +192,7 @@ fn roll_complex() {
                 assert(new_block_number1 == 789, 'Wrong block number #7');
                 assert(new_block_number2 == old_block_number2, 'Wrong block number #8');
             }
-        "
+        "#
         ),
         Contract::from_code_path(
             "RollChecker".to_string(),