+++ CHANGELOG.md
@@ -7,7 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Forge
+
+#### Added
+
+- result of the call to the trace
+
+#### Changed
+
+- `var()`, `read_txt()`, `read_json()`, `FileTrait::new()`, `declare()` now use regular strings (`ByteArray`) instead of short strings (`felt252`)
+
 ### Cast
+
+#### Changed
+
+- `declare()` now use regular strings (`ByteArray`) instead of short strings (`felt252`)
 
 #### Removed